clash vdw values from probe program for clash constraints + fix nonbond bugs

+++ LinearOptimizer/get_cross_conf_nonbonds_wrapper.py
@@ -1,16 +1,19 @@
 import subprocess
-import os
+import os, sys, pathlib
+sys.path.append(str(pathlib.Path(__file__).parent.parent))
 from UntangleFunctions import UNTANGLER_WORKING_DIRECTORY
 # Interface for phenix.python call
-def get_cross_conf_nonbonds(pdb_file_path):
+def get_cross_conf_nonbonds(pdb_file_path,use_previous=False,nonbonds=False):
 
-    tmp_file_path="tmp_nonbond_data.txt"
-    if os.path.exists(tmp_file_path):
-        os.remove(tmp_file_path)
-    script_path=os.path.join(UNTANGLER_WORKING_DIRECTORY,"PhenixEnvScripts","cross_conformation_nonbonds.py")
-    args=["phenix.python",script_path,pdb_file_path,tmp_file_path]
-    print (f"$Running {args}")
-    subprocess.run(args)#,stdout=log)
+    kind='nonbonds' if nonbonds else 'clashes'
+    tmp_file_path=f"tmp_{kind}_data.txt"
+    if not use_previous:
+        if os.path.exists(tmp_file_path):
+            os.remove(tmp_file_path)
+        script_path=os.path.join(UNTANGLER_WORKING_DIRECTORY,"PhenixEnvScripts",f"cross_conformation_{kind}.py")
+        args=["phenix.python",script_path,pdb_file_path,tmp_file_path]
+        print (f"$Running {args}")
+        subprocess.run(args)#,stdout=log)
 
 
     out_items = []
@@ -20,5 +23,8 @@ def get_cross_conf_nonbonds(pdb_file_path):
             vdw = float(vdw)
             is_symm = (is_symm.lower() == "true")
             out_items.append((pdb1,pdb2,vdw,is_symm))
-    os.remove(tmp_file_path)
+    #os.remove(tmp_file_path)
     return out_items
+
+if __name__ == "__main__":
+    get_cross_conf_nonbonds(*sys.argv[1:])
